file modify

+++ libft/ft_printf/ft_itoa_base.c
@@ -36,7 +36,7 @@ char			*ft_itoa_base(int value, int base)
 	res = NULL;
 	if (base == 10)
 		return (res = ft_itoa(value));
-	value = (value < 0) ? (unsigned int)value : value;
+	value = (value < 0) ? (unsigned int)value : (unsigned int)value;
 	i = ft_count(value, base);
 	res = (char*)malloc(sizeof(char) * i);
 	res[i] = 0;
