+++ libft/ft_printf/ft_ltoa_base_maj.c
@@ -35,7 +35,7 @@ char			*ft_ltoa_base_maj(long value, int base)
 	res = NULL;
 	if (base == 10)
 		return (ft_ltoa(value));
-	value = (value < 0) ? (unsigned long)value : value;
+	value = (value < 0) ? (unsigned long)value : (unsigned long)value;
 	i = ft_count(value, base);
 	res = (char*)malloc(sizeof(char) * i);
 	res[i] = 0;
